% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtimeforzx.R
\name{timeforzx}
\alias{timeforzx}
\title{Time series prediction}
\usage{
timeforzx(timeseries, mons)
}
\arguments{
\item{timeseries}{the time series data model}

\item{mons}{month}
}
\value{
Number of cases in a given month
}
\description{
Predict the number of cases in a given month in the future
}
\examples{
timeforzx(tsa,6)





}
