% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtimeplotzx.R
\name{timeplotzx}
\alias{timeplotzx}
\title{Plot of Time series prediction}
\usage{
timeplotzx(timeseries, mons)
}
\arguments{
\item{timeseries}{the time series data model}

\item{mons}{month}
}
\value{
A plot
}
\description{
A plot of the number of cases for a given month in the future
}
\examples{
timeplotzx(tsa,6)




}
