% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timezx.R
\name{timezx}
\alias{timezx}
\title{Time series analysis data preparation}
\usage{
timezx(infect, years, months)
}
\arguments{
\item{infect}{Infectious disease report card data.}

\item{years}{Starting year.}

\item{months}{Starting month.}
}
\value{
The time series data model
}
\description{
The data of infectious disease report card were preprocessed,
and then the time series data model was made.
}
\examples{
timezx(data,2012,1)


}
